refactor: RacingGame 클래스에서 Player가 해야 하는 일을 Player Class로 분리
